ルートと/oldページの入れ替えを実装
- RetroPageコンポーネントを作成し、レトロデザインのロジックを分離
- ModernPageコンポーネントを作成し、モダンデザインのロジックを分離
- index.jsx: RetroPageを呼び出しルートでレトロデザインを表示
- old.jsx: ModernPageを呼び出し/oldでモダンデザインを表示
- タイトル調整: ルート「オーガスタロウの館」、/old「オーガスタロウの館 - モダンver」
